requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
streamlit>=1.25.0
pymongo>=4.5.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pandas>=1.5.0
seaborn>=0.12.0
shap>=0.41.0
fpdf>=1.7.2
fastapi>=0.100.0
uvicorn>=0.23.0
pydicom>=2.4.0
opencv-python>=4.8.0
python-multipart